refactor: переместить компонент Header в виджеты и обновить стили, удалить неиспользуемые файлы
