+++ frontend/src/pages/ArticleDetails/ui/ArticleDetails.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import axios from "axios";
-import Header from "../../../components/header/Header.tsx";
+import Header from "../../../widgets/header";
 import myimg from "../../../assets/img_2.png";
 import myimg2 from "../../../assets/img_3.png";
 import "./ArticleDetails.css";
