+++ frontend/src/pages/CatsPage/ui/CatsPage.tsx
@@ -1,6 +1,6 @@
 // CatsPage.tsx
 import "./CatsPage.css";
-import Header from "../../../components/header/Header.tsx";
+import Header from "../../../widgets/header";
 import { useEffect, useState } from "react";
 import axios from "axios";
 import MiniBublik from "../../../components/miniBublikCard/miniBublik";
