+++ frontend/src/shared/assets/personLogo.png
